chore: removed Stack and Tree from tests and added them in ts files
