+++ packages/react/components/src/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import ReactDOM from 'react-dom';
-
+// TEST
 
 function SampleComponent() {
   const [msg, updateMsg] = useState('test');
@@ -9,8 +9,8 @@ function SampleComponent() {
     updateMsg('show something different');
   }, 10000);
   return <p>
-      <span>{msg}</span>
-    </p>;
+    <span>{msg}</span>
+  </p>;
 }
 
 ReactDOM.render(
